Safari-611.1.10/LayoutTests/media/media-source/media-source-istypesupported-case-sensitive-expected.txt


EXPECTED (MediaSource.isTypeSupported("video/mock; codecs=mock") == 'true') OK
EXPECTED (MediaSource.isTypeSupported("video/MOCK; codecs=mock") == 'true') OK
EXPECTED (MediaSource.isTypeSupported("video/mock; codecs=MOCK") == 'false') OK
EXPECTED (MediaSource.isTypeSupported("video/MOCK; codecs=MOCK") == 'false') OK
END OF TEST
